### 2017-07-27 #337 @pjotrp
It could be perceived I have a conflict of interest here. @biorelated can you be the editor of this Ruby submission?
