### 2017-07-27 #337 @biorelated
@krother, @amoeba would you be willing and happy to review this work?
