### 2017-07-28 #337 @pjotrp
Thanks @biorelated. Would be good to get one of the Ruby crowd. How about asking one of the users https://github.com/fstrozzi/bioruby-pipengine/stargazers?
