### 2017-07-31 #341 @whedon
[REVIEW]: Pipengine
**Submitting author:** @fstrozzi (<a href="http://orcid.org/0000-0002-6845-6982">Francesco Strozzi</a>)
**Repository:** <a href="https://github.com/fstrozzi/bioruby-pipengine" target ="_blank">https://github.com/fstrozzi/bioruby-pipengine</a>
**Version:** 0.9.3
