### 2017-08-02 4fb611f
Update paper.md
