### 2017-08-02 0587327
Update paper.md
