### 2017-08-02 a8e4822
Update paper.md
