+++ paper.md
@@ -23,7 +23,7 @@ date: 25 July 2017
 
 This is an ultra simple YAML-based pipeline execution engine. The tool allows defining a pipeline template in YAML, specifing command lines, resources and software to be used along with pipeline steps dependencies. Pipengine is a sample-centric tool, so the pipeline can then be applied over a single sample or multiple samples data, generating actual runnable bash scripts which can then be sumitted automatically to a scheduling system or runned locally. 
 
-The bash scripts generated by Pipengine includes error controls and logging for each step, plus the automatical generation of directories based on sample and pipeline steps names, the moving of input and output data across original and temporary folders if needed and a simple checkpoint strategy to avoid re-running already completed steps in a pipeline. This allows the user to avoid rewriting boiler plate code to perform all these necessary accessory tasks. 
+The bash scripts generated by Pipengine includes error controls and logging for each step, plus the automatical generation of directories based on sample and pipeline steps names, the moving of input and output data across original and temporary folders if needed, and a simple checkpoint strategy to avoid re-running already completed steps in a pipeline. This prevents the users to write boiler plate code to perform all these necessary accessory tasks. 
 
 Moreover Pipengine produces a stable and reproducible working and output tree for each analysis, which transparently stores all the results of each step of a pipeline for each sample analysed. In this way pipelines' intermediate or final results can be predictably accessed by the analysts and/or easily parsed with other tools. 
 
