### 2017-08-08 fstrozzi/bioruby-pipengine#40 @yannickwurm
Example data for testing?
is incomplete.

```sh
Yannick@n56-215 ~/g/b/t/data> pwd
/Users/Yannick/gitStuff/bioruby-pipengine/test/data
Yannick@n56-215 ~/g/b/t/data> tree -hlCF
.
├── [ 340]  mapping.yml
├── [1.2K]  pipeline-enh.yml
├── [1.5K]  pipeline.yml
├── [ 177]  samples.yml
└── [  68]  test/

1 directory, 4 files
Yannick@n56-215 ~/g/b/t/data> pipengine run -p pipeline.yml -f samples.yml -s mapping
I, [2017-08-08T16:07:38.059687 #52908]  INFO -- : Directory ./test/sampleA/trim not found
E, [2017-08-08T16:07:38.059818 #52908] ERROR -- : Found an unsubstituted tag <flowcell>\tLB:sampleA\tPL:ILLUMINA\tPU:<flowcell> . Terminating the execution
```
- perhaps this could be completed,  or an alternate could be provided (that does depend on  samtools/bwa installation)
