### 2017-08-08 fstrozzi/bioruby-pipengine#41 @yannickwurm
Specific dependencies?
Hi,
gem install worked seamlessly for me.
It is probably worth mentioning which ruby version(s) is/are required or compatible with this.
