### 2017-08-08 fstrozzi/bioruby-pipengine#41 @yannickwurm
(I also suppose that the PBS/Torque support was developed for a particular version - also worth mentioning)
