### 2017-08-08 fstrozzi/bioruby-pipengine#42 @yannickwurm
Idiotproofing
Running `pipengine` with nothing -> would be great if it pointed me to -h
(and maybe indicated version)
