### 2017-08-08 #341 @yannickwurm
## Paper:

 - [ ] needs a grammar checker (e.g. from M.S. Word).
 - [ ] Some of the sentences could be shortened to increase clarity
 - [ ] First sentence of Paragraph 2 is a long list - using bullet-points would make it more legible.
 - [ ] Some readers may not know what "pre-CWL/WDL era" means. I suggest defining any abbreviations that might be ambiguous and linking to suitable sites/refs.
 - [ ] Has no references. I don't know if that is a problem. Relevant refs could be to (PBS/Torque, Bioruby, etc).

## Readme:

 - [ ] A sentence in the readme could indicate why this was needed (e.g., why alternatives (pure bash, rake, whatever) were /are inappropriate). --> the JOSS review form requests this to be crystalclear in the paper as well (see "*A statement of need:*").
 - [ ] Community Guidelines: Please add 2-3 sentences on e.g. how someone would add support for another scheculder

"References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?"  @biorelated I don't know what this means. I think wer'e ok

## Code/behavior:

 - [ ] code includes a bunch of commented things
 - [ ] Test data: https://github.com/fstrozzi/bioruby-pipengine/issues/40
 - [ ] Versions of dependencies: https://github.com/fstrozzi/bioruby-pipengine/issues/41
 - [ ] default output: https://github.com/fstrozzi/bioruby-pipengine/issues/42
 - [ ] What are the other development branches for? Can they be deleted?
 - [ ] Shouldn't gemfile be more specific regarding version numbers to ensure that things run as expected/tested?
