### 2017-08-08 #341 @biorelated
Thank you very much @yannickwurm for taking your time to review this manuscript and code submission.  @fstrozzi please could you fix the or comment on the issues or concerns that have been raised by Yannick. In addition there are a couple of issues and checkboxes regarding the functionality and documentation of the software that may require some attention.
Thank you.
George
