+++ joss/paper.md
@@ -1,0 +1,49 @@
+---
+title: 'Pipengine: an ultra simple YAML-based pipeline execution engine'
+tags:
+  - pipeline
+  - workflows
+  - reproducibility
+authors:
+ - name: Francesco Strozzi
+   orcid: 0000-0002-6845-6982
+   affiliation: 1
+ - name: Raoul Jean Pierre Bonnal
+   orcid: 0000-0002-2123-6536
+   affiliation: 2
+affiliations:
+ - name: Enterome Bioscience, Paris - France
+   index: 1
+ - name: INGM - Isitituto Nazionale Genetica Molecolare "Romeo ed Enrica Invernizzi": Milan, Italy
+   index: 2
+date: 25 July 2017
+bibliography: paper.bib
+---
+
+# Summary
+
+This is an ultra simple YAML-based pipeline execution engine. The tool allows defining a pipeline template in YAML, specifying command lines, resources and software to be used along with pipeline steps dependencies. Pipengine is a sample-centric tool, so the pipeline can then be applied over a single sample or multiple samples data, generating actual runnable bash scripts which can then be submitted automatically to a scheduling system or run locally.
+
+The bash scripts generated by Pipengine includes a list of features such as: 
+
+* error controls and logging for each step
+
+* the automated generation of directories based on sample and pipeline steps names 
+
+* the moving of input and output data across original and temporary folders if needed 
+
+* a simple checkpoint strategy to avoid re-running already completed steps in a pipeline.
+
+All these features prevent the users to write boiler plate code to perform all these necessary accessory tasks.
+
+Moreover, Pipengine produces a stable and reproducible working and output tree for each analysis, which transparently stores all the results of each step of a pipeline for each sample analyzed. In this way pipelines' intermediate or final results can be predictably accessed by the analysts and/or easily parsed with other tools.
+
+The software was developed back in 2012, when more generalized schemas such as for instance the Common Workflow Language were not yet defined, and thus was among the firsts utilities to introduce the concept of using simple YAML as a template format to define reusable bioinformatics pipelines. 
+
+Pipengine has been used across several research groups and bioinformatics core facilities since its first appearance. It directly supports the PBS/Torque scheduler for submission of jobs, but given that the support for a scheduler is based on specific options written automatically inside the bash scripts generated by the tool, it can be easily adapted to work with other schedulers, if needed.
+
+Pipengine is written in Ruby and is available for download as a BioRuby Gem.
+
+# References
+
+
