### 2017-08-12 458f0c3
changed example pipeline and added all tags in samples.yml. Addresses #40
