### 2017-08-12 0399fa1
Running pipengine with no options prompts to help screen #42
