### 2017-08-12 daaea15
added pipengine advantages and short explanation on how to extend the code
