### 2017-08-12 dffbe77
specified ruby version. Addresses #41
