### 2017-08-12 d7881f3
cleaned up commented and unused code
