### 2017-08-12 0c4868d
cleaned up Gemfile and specified Ruby version. #41
