### 2017-08-12 #341 @fstrozzi
Hi @biorelated and @yannickwurm ,
first of all thanks for your time in reviewing this submission and for providing useful feedbacks.

Yannick, here are the answers to your points.

**Paper**

A new version of the paper.md file has been uploaded addressing all the points raised, a paper.bib file was also added with the relevant references for Pipengine. Both are available in the "joss" folder.

**Readme**
The Readme file was improved, with a paragraph on what we considered the Pipengine advantages over previous approaches and the reason why it was developed. Along with that, a new section has been added at the end, briefly describing the source code organisation and where changes need to be made to extend Pipengine and to work with e.g. other job schedulers.

**Code**
Code has been cleaned up of old commented lines, the Gemfile now lists minimum version for the two gems Pipengine depends on and the minimum Ruby version.

The issues opened on the repository where also addressed with specific commits (I will let Yannick review those and close the issues if the changes will be considered satisfactory).

For other branches, if those pose a real problem for the archiving and publication on JOSS I will remove them from the repository, otherwise I would prefer to keep them, they contain development and experimental code that could turn out to be useful or picked up by others in the future.

A new version (0.9.6) is now available and pushed on RubyGems, which includes all the changes made to answer to reviewer's comments.

Thank you
Francesco
