### 2017-08-25 #341 @yannickwurm
Thanks v much - yes this looks much improved, but I must apologise that I am unable at this time to carefully doublecheck the authors work.

However, I do note three issues with the readme:
 1. "pipenengine"
 2. Executing the followign line (given in the readme) doesn't work:
```
Yannick@ybook ~/g/b/t/data> pipengine -p pipeline.yml -f samples.yml -s mapping

Launcher for Complex Biological Pipelines . Copyright(C) 2012 Francesco Strozzi, Raoul Jean Pierre Bonnal

PipEngine v0.9.6
List of available commands:
	run	Submit pipelines to the job scheduler

Error: unknown argument '-p'.
Try --help for help.
```

3. I also had the impression that running `pipengine run -d -p pipeline.yml -f samples.yml -s mapping` would create a script onto standard out (or into a local file?). I don't see any such file or output. This may be due to the fact that the `samples.yml` file includes absolute paths to files that don't exist on my system.
