+++ README.md
@@ -483,6 +483,8 @@ When multiple steps are run in the same job, by default PipEngine will generate 
 :: Examples ::
 ==============
 
+All these files can be found into the test/examples directory of the repository.
+
 Example 1: One step and multiple command lines
 ----------------------------------------------
 
@@ -529,18 +531,27 @@ pipengine run -p pipeline.yml -f samples.yml -s mapping -d
 will generate a runnable shell script for each sample:
 
 ```shell
-#!/bin/bash
-#PBS -N 37735f50-mapping
-#PBS -l ncpus=11
+#!/usr/bin/env bash
+#PBS -N 2c57c1a853-sampleA-mapping
+#PBS -d ./working
+#PBS -l nodes=1:ppn=11
+if [ ! -f ./working/sampleA/mapping/checkpoint ]
+then
+echo "mapping 2c57c1a853-sampleA-mapping start `whoami` `hostname` `pwd` `date`."
 
-mkdir -p /storage/results/sampleA/mapping
-cd /storage/results/sampleA/mapping
-ls /ngs_reads/sampleA/*_R1_*.gz | xargs zcat | pigz -p 10 >> R1.fastq.gz
-ls /ngs_reads/sampleA/*_R2_*.gz | xargs zcat | pigz -p 10 >> R2.fastq.gz
-/software/bwa sampe -P /genomes/bwa_index/genome <(/software/bwa aln -t 4 -q 20 /genomes/bwa_index/genome R1.fastq.gz) <(/software/bwa aln -t 4 -q 20 /genomes/bwa_index/genome R2.fastq.gz) R1.fastq.gz R2.fastq.gz | /software/samtools view -Sb - > sampleA.bam
-rm -f R1.fastq.gz R2.fastq.gz
+mkdir -p ./working/sampleA/mapping
+cd ./working/sampleA/mapping
+ls /ngs_reads/sampleA/*_R1_*.gz | xargs zcat | /software/pigz -p 10 >> R1.fastq.gz || { echo "mapping 2c57c1a853-sampleA-mapping FAILED 0 `whoami` `hostname` `pwd` `date`."; exit 1; }
+ls /ngs_reads/sampleA/*_R2_*.gz | xargs zcat | /software/pigz -p 10 >> R2.fastq.gz || { echo "mapping 2c57c1a853-sampleA-mapping FAILED 1 `whoami` `hostname` `pwd` `date`."; exit 1; }
+/software/bwa sampe -P /storage/genomes/bwa_index/genome <(/software/bwa aln -t 4 -q 20 /storage/genomes/bwa_index/genome R1.fastq.gz) <(/software/bwa aln -t 4 -q 20 /storage/genomes/bwa_index/genome R2.fastq.gz) R1.fastq.gz R2.fastq.gz | /software/samtools view -Sb - > sampleA.bam || { echo "mapping 2c57c1a853-sampleA-mapping FAILED 2 `whoami` `hostname` `pwd` `date`."; exit 1; }
+rm -f R1.fastq.gz R2.fastq.gz || { echo "mapping 2c57c1a853-sampleA-mapping FAILED 3 `whoami` `hostname` `pwd` `date`."; exit 1; }
+echo "mapping 2c57c1a853-sampleA-mapping finished `whoami` `hostname` `pwd` `date`."
+touch ./working/sampleA/mapping/checkpoint
+else
+echo "mapping 2c57c1a853-sampleA-mapping already executed, skipping this step `whoami` `hostname` `pwd` `date`."
+fi
 ```
[…]
