### 2017-08-26 2e9d8be
included examples files from Readme
