+++ test/examples/samples.yml
@@ -1,0 +1,10 @@
+resources:
+  index: /storage/genomes/bwa_index/genome
+  genome: /storage/genomes/genome.fa
+  output: ./working
+
+samples:
+  sampleA: /ngs_reads/sampleA
+  sampleB: /ngs_reads/sampleB
+  sampleC: /ngs_reads/sampleC
+  sampleD: /ngs_reads/sampleD
