### 2017-08-26 5073e9c
fixed cpu number in the example pipeline
