### 2017-08-26 #341 @fstrozzi
Hi Yannick,
thanks for your checks, the typos in the Readme command lines were corrected.

To answer to point number 3, the test files into test/data do work as expected, the runnable scripts are created into an output directory that is specified in the samples.yml file, which in this case is a local ./test directory. I have made this behaviour clearer in the "What happens at run time" section of the Readme.

I have also updated the "Examples" section of the Readme, and the same files described there are also available in the test/examples folder so that the command lines can be tested directly. The output scripts will be generated into a local ./working directory in the case of the examples file.

Please let me know if you encounter other issues with the tests or documentation so that we can address those.

Thank you very much.
Francesco
