### 2017-08-26 #341 @fstrozzi
Thanks Yannick,
if (or when) everything is good for you, could you please let it know to the editor (George) ?

Thanks and cheers
Francesco
