### 2017-08-28 #341 @fstrozzi
Thanks Yannick ! (@yannickwurm) .

George (@biorelated), I think we have finished here. What I need to do next ? I have created and pushed a new gem version (0.9.7) which includes all the changes made during the review process. I think the submission needs to be changed to reflect the last version of the package, but I don't know if I can do this or if this is up to you.

Thanks
Francesco
