### 2017-08-28 58f42e5
Update paper.md
