### 2017-08-28 f4d8cb4
Update paper.md
