+++ joss/paper.md
@@ -36,7 +36,7 @@ The bash scripts generated by Pipengine includes a list of features such as:
 
 All these features prevent the users to write boiler plate code to perform all these necessary accessory tasks.
 
-Moreover, Pipengine produces a stable and reproducible working and output tree for each analysis, which transparently stores all the results of each step of a pipeline for each sample analyzed. In this way pipelines' intermediate or final results can be predictably accessed by the analysts and/or easily parsed with other tools.
+Moreover, Pipengine creates a stable and reproducible working and output tree for each analysis, which transparently stores all the results of each step of a pipeline for each sample analyzed. In this way pipelines' intermediate or final results can be predictably accessed by the analysts and/or easily parsed with other tools.
 
 The software was developed back in 2012, when more generalized schemas such as for instance the Common Workflow Language were not yet defined, and thus was among the firsts utilities to introduce the concept of using simple YAML as a template format to define reusable bioinformatics pipelines. 
 
