### 2017-08-28 #341 @biorelated
Thank you very much Yannick (@yannickwurm)  for providing a detailed review. The submission will reflect the latest version of the package. Gem version (0.9.7) at the time of writing.

@arfon this work is ready for publishing.
