### 2017-08-29 #341 @arfon
@fstrozzi - Your references (in `paper.bib`) currently aren't being cited in the paper body so they won't be compiled by Pandoc when we process the paper. Could you cite them directly please? You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
