### 2017-08-29 507e53e
Update paper.bib
