+++ joss/paper.bib
@@ -1,11 +1,11 @@
-@online{
+@online{Torque2017,
 	author = "Adaptive Computing Inc.",
   title = {TORQUE Resource Manager},
 	year = {2017},
 	url = {http://www.adaptivecomputing.com/products/open-source/torque/}
 }
 
-@online{
+@online{CWL2017,
 	author = "Common Workflow Language",
   title = {Common Workflow Language},
 	year = {2017},
@@ -13,7 +13,7 @@
 	doi = {dx.doi.org/10.6084/m9.figshare.3115156.v2}
 }
 
-@article{
+@article{Goto2010,
 	title = {Bioinformatics software for the Ruby programming language},
 	author = {Goto N, Prins P, Nakao M, Bonnal RJP, Aerts J, Katayama T},
 	year = {2010},
@@ -22,7 +22,7 @@
 }
 
 
-@article{
+@article{Bonnal2010,
 	title = {Biogem: An effective tool-based approach for scaling up open source software development in bioinformatics},
 	author = {Bonnal RJP, Aerts J, Githinji G, Goto N, Maclean D, Miller CA, Mishima H, Pagani M, Ramirez-gonzalez R, Smant G, Strozzi F, Syme R, Vos R, Wennblom TJ, Woodcroft BJ, Katayama T, Prins P},
 	year = {2012},
