### 2017-08-29 7cc78af
Update paper.md
