+++ joss/paper.md
@@ -38,11 +38,11 @@ All these features prevent the users to write boiler plate code to perform all t
 
 Moreover, Pipengine creates a stable and reproducible working and output tree for each analysis, which transparently stores all the results of each step of a pipeline for each sample analyzed. In this way pipelines' intermediate or final results can be predictably accessed by the analysts and/or easily parsed with other tools.
 
-The software was developed back in 2012, when more generalized schemas such as for instance the Common Workflow Language were not yet defined, and thus was among the firsts utilities to introduce the concept of using simple YAML as a template format to define reusable bioinformatics pipelines. 
+The software was developed back in 2012, when more generalized schemas such as for instance the Common Workflow Language [@CWL2017] were not yet defined, and thus was among the firsts utilities to introduce the concept of using simple YAML as a template format to define reusable bioinformatics pipelines. 
 
-Pipengine has been used across several research groups and bioinformatics core facilities since its first appearance. It directly supports the PBS/Torque scheduler for submission of jobs, but given that the support for a scheduler is based on specific options written automatically inside the bash scripts generated by the tool, it can be easily adapted to work with other schedulers, if needed.
+Pipengine has been used across several research groups and bioinformatics core facilities since its first appearance. It directly supports the PBS/Torque scheduler [@Torque2017] for submission of jobs, but given that the support for a scheduler is based on specific options written automatically inside the bash scripts generated by the tool, it can be easily adapted to work with other schedulers, if needed.
 
-Pipengine is written in Ruby and is available for download as a BioRuby Gem.
+Pipengine is written in Ruby and is available for download as a BioRuby Gem [@Goto2010, @Bonnal2012].
 
 # References
 
