### 2017-08-29 c1011ab
Update paper.md
