+++ joss/paper.md
@@ -42,7 +42,7 @@ The software was developed back in 2012, when more generalized schemas such as f
 
 Pipengine has been used across several research groups and bioinformatics core facilities since its first appearance. It directly supports the PBS/Torque scheduler [@Torque2017] for submission of jobs, but given that the support for a scheduler is based on specific options written automatically inside the bash scripts generated by the tool, it can be easily adapted to work with other schedulers, if needed.
 
-Pipengine is written in Ruby and is available for download as a BioRuby Gem [@Goto2010, @Bonnal2012].
+Pipengine is written in Ruby and is available for download as a BioRuby Gem [@Goto2010; @Bonnal2012].
 
 # References
 
