+++ joss/paper.bib
@@ -26,6 +26,6 @@
 	title = {Biogem: An effective tool-based approach for scaling up open source software development in bioinformatics},
 	author = {Bonnal RJP, Aerts J, Githinji G, Goto N, Maclean D, Miller CA, Mishima H, Pagani M, Ramirez-gonzalez R, Smant G, Strozzi F, Syme R, Vos R, Wennblom TJ, Woodcroft BJ, Katayama T, Prins P},
 	year = {2012},
-	doi = {doi.org/10.1093/bioinformatics/btq475},
+	doi = {doi.org/10.1093/bioinformatics/bts080},
 	journal = {Bioinformatics}
 }
