### 2017-08-29 #341 @fstrozzi
@arfon I've updated the bib file and added direct citations into the paper body. It should be fine now, thanks.
