### 2017-08-29 b52d041
Update paper.md
