+++ joss/paper.bib
@@ -9,7 +9,7 @@
 	author = "Common Workflow Language",
   title = {Common Workflow Language},
 	year = {2017},
-	url = {http://www.commonwl.org}
+	url = {http://www.commonwl.org},
 	doi = {dx.doi.org/10.6084/m9.figshare.3115156.v2}
 }
 
