### 2017-08-29 #341 @arfon
Thanks @fstrozzi. Could you please merge this PR https://github.com/fstrozzi/bioruby-pipengine/pull/43

Following that, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
