### 2017-08-30 #341 @fstrozzi
@arfon done, and here is the DOI from Zenodo: https://doi.org/10.5281/zenodo.851891 .

Thanks !
